\name{Example_run}
\alias{Example_run}
\docType{data}
\title{Example Luminex run file.}
\description{This is an exemplary Luminex run file provided here for testing purpose and as
    a "proof of concept". It can be loaded into R using \code{\link{read.multiplex}} function.
    }
\format{
  This is an output file from Luminex IS software. To review it's internal structure
  please open it in a text editor or a spreadsheet.  
}
\details{
  This output file is provided with the \code{\link{immunoassay}} package as an example of
  a loadable Luminex output file.
}
\examples{\dontrun{
    read.multiplex("your-path-here/Example_run.csv")
}}
\keyword{ datasets }
\keyword{ immunoassay }
\keyword{ Luminex }
\keyword{ raw data }
\keyword{ csv }

