\name{Luminex-internal}
\alias{sigfit.auto}
\alias{sigfit.default}
\alias{check.sigfit}
\alias{check.ima}
\alias{cv}
\alias{mesdci}
\title{Internal 'immunoassay' library objects}
\description{Internal 'immunoassay' library objects.}
\details{These are not to be called by the user.}
\keyword{internal}
\keyword{immunoassay}